Move AdsController to package folder
