+++ core/ui/src/commonMain/kotlin/org/pointyware/xyz/ui/ads/AdsController.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Pointyware. Use of this software is governed by the GPL-3.0 license.
  */
 
-package org.pointyware.xyz.core.ui.ads
+package org.pointyware.xyz.ui.ads
 
 /**
  * Abstracts Ad logic.
